move Simcdl to device dir
